requirementsAct.txt
alembic==1.11.0
asgiref==3.5.2
banal==1.0.6
cached-property==1.5.2
dataclasses==0.6
dataset==1.5.2
debugpy==1.6.2
distlib==0.3.5
Django==4.0.6
django-bootstrap-form==3.4
django-contrib-comments==2.2.0
django-models==0.5.1
django-object-tools==2.0.0
djangocms-admin-style==3.1.1
djangorestframework==3.14.0
filelock==3.4.2
Mako==1.2.4
MarkupSafe==2.1.2
platformdirs==2.5.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.13
sqlparse==0.4.2
typing_extensions==4.5.0
tzdata==2022.1
Unipath==1.1
virtualenv==20.15.1
virtualenv-clone==0.5.7
